+++ hbw/config/routes.rb
@@ -1,14 +1,14 @@
 HBW::Engine.routes.draw do
   defaults format: :json do
-    resources :tasks, only: [:index] do
+    get 'tasks/claiming', to: 'tasks#claiming'
+    get 'tasks/list', to: 'tasks#list'
+    get 'tasks/count', to: 'tasks#count'
+
+    resources :tasks, only: [:index, :show] do
       put :form, on: :member, action: :submit
       get :lookup, on: :member
       post :claim, on: :member, action: :claim
     end
-
-    get 'tasks/claiming', to: 'tasks#claiming'
-    get 'tasks/list', to: 'tasks#list'
-    get 'tasks/count', to: 'tasks#count'
 
     resources :buttons, only: [:index, :create]